`@remotion/studio`: Extract fullscreen state hook
